renames scripts/common to scripts/base to match scss directory names
