+++ assets/scripts/base/detect.js
