+++ assets/scripts/base/helpers.js
