+++ assets/scripts/base/index.js
